+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## 2.2 branch
 ### 2.2.1
 * fixed to work under Windows;
+* added `Driver::getConfig()` method, removed `Driver::$config` property. This
+    allows you to get the configuration values safely;
 * fixed a bug for export and import executables with Postgres databases;
 * before importing a sqlite backup, each table is dropped, rather than deleting
     the database file;